requirements.txt
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2
xgboost>=1.7
joblib
matplotlib
seaborn
streamlit
pillow
pytesseract
opencv-python
easyocr
tqdm
